+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - New user *can* receive the free dollar promotion after verification
 - Newly created messages will be automatically selected by default
 - Workflow duration in seconds
+- Auto categorisation before prompt is sent
+  - Allows user to pick category, display which category will be selected in advance
+    and reduce request durations.
 - Edit internet search instructions in settings.
 
 ### Changed